+++ src/shapes/Object/InteractiveObject.ts
@@ -37,13 +37,6 @@ export type TStyleOverride = ControlRenderingStyleOverride &
     }
   >;
 
-export interface DragMethods {
-  shouldStartDragging(): boolean;
-  onDragStart(e: DragEvent): boolean;
-}
-
-export type FabricObjectWithDragSupport = InteractiveFabricObject & DragMethods;
-
 const interactiveDefaults = {
   noScaleCache: true,
   centeredScaling: false,
@@ -125,13 +118,10 @@ export class InteractiveFabricObject<
   declare moveCursor: CSSStyleDeclaration['cursor'] | null;
 
   /**
-   * Describe object's corner position in canvas element coordinates.
-   * properties are depending on control keys and padding the main controls.
-   * each property is an object with x, y and corner.
-   * The `corner` property contains in a similar manner the 4 points of the
-   * interactive area of the corner.
-   * The coordinates depends from the controls positionHandler and are used
-   * to draw and locate controls
+   * The object's controls' position in viewport coordinates
+   * Calculated by {@link Control#positionHandler} and {@link Control#calcCornerCoords}, depending on {@link padding}.
+   * `corner/touchCorner` describe the 4 points forming the interactive area of the corner.
+   * Used to draw and locate controls.
    */
   declare oCoords: Record<string, TOCoord>;
 
@@ -330,16 +320,12 @@ export class InteractiveFabricObject<
   }
 
   /**
-   * Sets corner and controls position coordinates based on current angle, width and height, left and top.
-   * oCoords are used to find the corners
-   * aCoords are used to quickly find an object on the canvas
-   * lineCoords are used to quickly find object during pointer events.
+   * @override set controls' coordinates as well
    * See {@link https://github.com/fabricjs/fabric.js/wiki/When-to-call-setCoords} and {@link http://fabricjs.com/fabric-gotchas}
    * @return {void}
    */
   setCoords(): void {
     super.setCoords();
-    // set coordinates of the draggable boxes in the corners used to scale/rotate the image
     this.canvas && (this.oCoords = this.calcOCoords());
   }
 
@@ -661,11 +647,28 @@ export class InteractiveFabricObject<
   }
 
   /**
+   * Override to customize Drag behavior
+   * Fired from {@link Canvas#_onMouseMove}
+   * @returns true in order for the window to start a drag session
+   */
+  shouldStartDragging() {
+    return false;
+  }
+
+  /**
+   * Override to customize Drag behavior\
+   * Fired once a drag session has started
+   * @returns true to handle the drag event
+   */
+  onDragStart(e: DragEvent) {
+    return false;
+  }
+
+  /**
    * Override to customize drag and drop behavior
-   * return true if the object currently dragged can be dropped on the target
    * @public
    * @param {DragEvent} e
-   * @returns {boolean}
+   * @returns {boolean} true if the object currently dragged can be dropped on the target
    */
   canDrop(e: DragEvent): boolean {
     return false;
@@ -677,7 +680,6 @@ export class InteractiveFabricObject<
    * example: render the selection status for the part of text that is being dragged from a text object
    * @public
    * @param {DragEvent} e
-   * @returns {boolean}
    */
   renderDragSourceEffect(e: DragEvent) {
     // for subclasses
@@ -690,7 +692,6 @@ export class InteractiveFabricObject<
    * object will change when dropping. example: show the cursor where the text is about to be dropped
    * @public
    * @param {DragEvent} e
-   * @returns {boolean}
    */
   renderDropTargetEffect(e: DragEvent) {
     // for subclasses